implemented StorageManager.

+++ src/rstar/dto/PointDTO.java
@@ -7,9 +7,9 @@ package rstar.dto;
  */
 public class PointDTO extends AbstractDTO{
     public float oid;
-    public double[] coords;
+    public float[] coords;
 
-    public PointDTO(float oid, double[] coords) {
+    public PointDTO(float oid, float[] coords) {
         this.oid = oid;
         this.coords = coords;
     }
